Upgrade dependencies v2024.1.0 (#731)
* Reporter to ReportNode in java part


* Action + double for limit reduction


* Dynawo


* nad coordinates


* Fixes get limit reduction returns double


* Update pom


* Fix loadflow parameter test


* test


* write slack bus default value is now true (in docs)


* Deprecate reporter


* Report node in documentation


* fix network modifications


* Add tests


* Add tests network modifications


* Fix


* Fix


* Define constants


---------

+++ java/src/main/java/com/powsybl/python/commons/PyPowsyblApiHeader.java
@@ -507,10 +507,10 @@ public final class PyPowsyblApiHeader {
         void setAcceptableDuration(int acceptableDuration);
 
         @CField("limit_reduction")
-        float getLimitReduction();
+        double getLimitReduction();
 
         @CField("limit_reduction")
-        void setLimitReduction(float limitReduction);
+        void setLimitReduction(double limitReduction);
 
         @CField("value")
         double getValue();
